pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "spec-orch"
version = "0.1.0"
description = "SpecOrch MVP prototype"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "typer>=0.24,<0.25",
  "httpx>=0.28",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9,<10",
  "ruff>=0.9",
  "mypy>=1.14",
  "types-PyYAML>=6.0",
]

[project.scripts]
spec-orch = "spec_orch.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/spec_orch/cli.py" = ["B008"]
"src/spec_orch/services/_archived/*.py" = ["E501", "SIM102"]
"src/spec_orch/services/codex_exec_builder_adapter.py" = ["E501"]
"tests/**" = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "spec_orch.services._archived.codex_harness_builder_adapter"
ignore_errors = true
